Update to v1.1.7

+++ lib/xmlconv/util/polling_manager.rb
@@ -173,7 +173,11 @@ module XmlConv
 			def initialize(system)
 				@system = system
 			end
-			def load_sources(&block)
+      def load_sources(&block)
+        unless File.exist?(CONFIG.polling_file)
+          SBSM.warn("Could not find #{CONFIG.polling_file}. Skip polling")
+          return
+        end
 				file = File.open(CONFIG.polling_file)
 				YAML.load_documents(file) { |mission|
 					block.call(mission)
@@ -182,8 +186,10 @@ module XmlConv
 				file.close if(file)
 			end
 			def poll_sources
+        @@showed_content ||= false
 				load_sources do |source|
           begin
+            SBSM.info "PollingManage has source #{source.inspect}" unless @@showed_content
             source.poll { |transaction|
               @system.execute(transaction)
             }
@@ -197,6 +203,7 @@ module XmlConv
             Util::Mail.notify source.error_recipients, subject, body
           end
         end
+        @@showed_content = true
 			end
 		end
 	end
